feat(coreaudio): move to different backends + add extension to get audio unit + prelude modules + warnings fix
